requirements.txt
airtable-python-wrapper==0.8.0
appnope==0.1.0
asgiref==3.2.3
backcall==0.1.0
branca==0.3.1
certifi==2018.1.18
chardet==3.0.4
Click==7.0
decorator==4.4.0
Django==3.0.3
django-crispy-forms==1.8.1
Flask==1.1.1
folium==0.10.0
glob2==0.7
idna==2.6
ipython==7.7.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
Markdown==3.2.1
MarkupSafe==1.1.1
mistune==0.8.4
numpy==1.17.0
pandas==0.25.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2017.3
requests==2.23.0
six==1.11.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.25.8
virtualenv==16.7.4
wcwidth==0.1.7
Werkzeug==0.15.5
xlrd==1.2.0
